+++ db/migrate/20150305185613_add_information_to_datasets.rb
@@ -1,0 +1,10 @@
+class AddInformationToDatasets < ActiveRecord::Migration
+  def change
+    add_column :datasets, :info_what, :string
+    add_column :datasets, :info_who, :text
+    add_column :datasets, :info_when, :string
+    add_column :datasets, :info_funded_by, :text
+    add_column :datasets, :info_citation, :text
+    add_column :datasets, :info_size, :bigint, null: false, default: 0
+  end
+end