+++ CHANGELOG.md
@@ -35,6 +35,7 @@
     - download the full dataset using the NSRR gem
     - browse all covariates online
   - Dataset documentation pages now scale better with screen resolution to make text more legible
+  - Dataset owners can now highlight key information about their dataset
 - **Agreement Changes**
   - Step 2 now has additional instructions to clarify the importance of describing the "Specific Purpose", and selecting an appropriate number of datasets
   - The signature step now allows users to opt out of signing the agreement if they have institutional requirements forbidding signing of agreements